requirements.txt
botbuilder-core>=4.7.0
flask
flask-restful
